Move dask/array/top.py code to dask/blockwise.py
Now that dask.dataframe is also using this code we choose to rename it
away from top (tensor operation) to a more generic term, blockwise.

We also move it down to root level.

For now we keep the term dask.array.atop, though we may choose to modify
this in a future commit.
