+++ dask/array/reductions.py
@@ -12,7 +12,7 @@ from toolz import compose, partition_all, get, accumulate, pluck
 
 from . import chunk
 from .core import _concatenate2, Array, atop, handle_out
-from .top import lol_tuples
+from ..blockwise import lol_tuples
 from .creation import arange
 from .ufunc import sqrt
 from .utils import validate_axis
