+++ dask/dataframe/optimize.py
@@ -5,7 +5,7 @@ from ..optimization import cull, fuse_getitem, fuse
 from .. import config, core
 from ..highlevelgraph import HighLevelGraph
 from ..utils import ensure_dict
-from ..array.top import optimize_atop
+from ..blockwise import optimize_blockwise
 
 try:
     import fastparquet  # noqa: F401
@@ -16,7 +16,7 @@ except ImportError:
 def optimize(dsk, keys, **kwargs):
 
     if isinstance(dsk, HighLevelGraph):
-        dsk = optimize_atop(dsk, keys=list(core.flatten(keys)))
+        dsk = optimize_blockwise(dsk, keys=list(core.flatten(keys)))
 
     dsk = ensure_dict(dsk)
     from .io import dataframe_from_ctable
